Refactor and default pets apllication on / + pets record testing GET request
