+++ src/main/java/org/example/helloWorld/Hello.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.helloWorld;
 
 public class Hello {
