+++ src/main/java/org/example/helloWorld/HelloWorldResource.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.helloWorld;
 
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
