+++ src/main/java/org/example/pets/Pets.java
@@ -1,0 +1,3 @@
+package org.example.pets;
+
+public record Pets (String name) {}
