+++ src/main/java/org/example/pets/PetsApplication.java
@@ -1,0 +1,8 @@
+package org.example.pets;
+
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
+
+@ApplicationPath("/")
+public class PetsApplication extends Application {
+}
